Refactor: move nodule to src folder
